Fixed some more warnings

+++ src/TGUI/Widgets/SpinButton.cpp
@@ -45,7 +45,7 @@ namespace tgui
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-    SpinButton::Ptr SpinButton::create(int minimum, int maximum)
+    SpinButton::Ptr SpinButton::create(float minimum, float maximum)
     {
         auto spinButton = std::make_shared<SpinButton>();
 
@@ -393,13 +393,13 @@ namespace tgui
         Widget::load(node, renderers);
 
         if (node->propertyValuePairs["minimum"])
-            setMinimum(tgui::stoi(node->propertyValuePairs["minimum"]->value));
+            setMinimum(tgui::stof(node->propertyValuePairs["minimum"]->value));
         if (node->propertyValuePairs["maximum"])
-            setMaximum(tgui::stoi(node->propertyValuePairs["maximum"]->value));
+            setMaximum(tgui::stof(node->propertyValuePairs["maximum"]->value));
         if (node->propertyValuePairs["value"])
-            setValue(tgui::stoi(node->propertyValuePairs["value"]->value));
+            setValue(tgui::stof(node->propertyValuePairs["value"]->value));
         if (node->propertyValuePairs["step"])
-            setStep(tgui::stoi(node->propertyValuePairs["step"]->value));
+            setStep(tgui::stof(node->propertyValuePairs["step"]->value));
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
